third_party/blink/web_tests/external/wpt/trusted-types/trusted-types-reporting-for-Document-write-expected.txt

This is a testharness.js-based test.
[FAIL] No violation reported for write() with TrustedHTML arguments.
  assert_equals: no violation reported expected 0 but got 1
[FAIL] No violation reported for writeln() with TrustedHTML arguments.
  assert_equals: no violation reported expected 0 but got 1
Harness: the test ran to completion.
